+++ project/boot/src/boot.c
@@ -24,12 +24,12 @@ static struct ffs_area_desc boot_area_descs[] = {
 };
 
 /** Contains indices of the areas which can contain image data. */
-static uint16_t boot_img_areas[] = {
+static uint8_t boot_img_areas[] = {
     5, 6, 7, 8, 9, 10, 11,
 };
 
 /** Areas representing the beginning of image slots. */
-static uint16_t boot_slot_areas[] = {
+static uint8_t boot_slot_areas[] = {
     5, 8,
 };
